Use stack to build package
Convert the project to build using stack instead of only
cabal; more reproducible builds is a good thing.

Unfortunately, the mixins feature of cabal is not currently
compatible with some stack functionality, so we instead use
the -XNoImplicitPrelude extension and manually import relude
modules.

The package structure was also changed to not export a
library, but only an executable (renamed from "shaunwhite-exe"
to "shaunwhite").